Enum irsc::reply::Reply

pub enum Reply<'a> {
    RPL_WELCOME(CS<'a>),
    RPL_YOURHOST(CS<'a>),
    RPL_CREATED(CS<'a>),
    RPL_MYINFO(CS<'a>),
    RPL_BOUNCE(CS<'a>),
    RPL_USERHOST(CS<'a>),
    RPL_ISON(CS<'a>),
    RPL_AWAY(CS<'a>),
    RPL_UNAWAY(CS<'a>),
    RPL_NOWAWAY(CS<'a>),
    RPL_WHOISUSER(CS<'a>),
    RPL_WHOISSERVER(CS<'a>),
    RPL_WHOISOPERATOR(CS<'a>),
    RPL_WHOISIDLE(CS<'a>),
    RPL_ENDOFWHOIS(CS<'a>),
    RPL_WHOISCHANNELS(CS<'a>),
    RPL_WHOWASUSER(CS<'a>),
    RPL_ENDOFWHOWAS(CS<'a>),
    RPL_LISTSTART,
    RPL_LIST(CS<'a>),
    RPL_LISTEND(CS<'a>),
    RPL_UNIQOPIS(CS<'a>),
    RPL_CHANNELMODEIS(CS<'a>),
    RPL_NOTOPIC(CS<'a>),
    RPL_TOPIC(CS<'a>),
    RPL_INVITING(CS<'a>),
    RPL_SUMMONING(CS<'a>),
    RPL_INVITELIST(CS<'a>),
    RPL_ENDOFINVITELIST(CS<'a>),
    RPL_EXCEPTLIST(CS<'a>),
    RPL_ENDOFEXCEPTLIST(CS<'a>),
    RPL_VERSION(CS<'a>),
    RPL_WHOREPLY(CS<'a>),
    RPL_ENDOFWHO(CS<'a>),
    RPL_NAMREPLY(CS<'a>),
    RPL_ENDOFNAMES(CS<'a>),
    RPL_LINKS(CS<'a>),
    RPL_ENDOFLINKS(CS<'a>),
    RPL_BANLIST(CS<'a>),
    RPL_ENDOFBANLIST(CS<'a>),
    RPL_INFO(CS<'a>),
    RPL_ENDOFINFO(CS<'a>),
    RPL_MOTDSTART(CS<'a>),
    RPL_MOTD(CS<'a>),
    RPL_ENDOFMOTD(CS<'a>),
    RPL_YOUREOPER(CS<'a>),
    RPL_REHASHING(CS<'a>),
    RPL_YOURESERVICE(CS<'a>),
    RPL_TIME(CS<'a>),
    RPL_USERSSTART(CS<'a>),
    RPL_USERS(CS<'a>),
    RPL_ENDOFUSERS(CS<'a>),
    RPL_NOUSERS(CS<'a>),
    RPL_TRACELINK(CS<'a>),
    RPL_TRACECONNECTING(CS<'a>),
    RPL_TRACEHANDSHAKE(CS<'a>),
    RPL_TRACEUNKNOWN(CS<'a>),
    RPL_TRACEOPERATOR(CS<'a>),
    RPL_TRACEUSER(CS<'a>),
    RPL_TRACESERVER(CS<'a>),
    RPL_TRACESERVICE(CS<'a>),
    RPL_TRACENEWTYPE(CS<'a>),
    RPL_TRACECLASS(CS<'a>),
    RPL_TRACERECONNECT(CS<'a>),
    RPL_TRACELOG(CS<'a>),
    RPL_TRACEEND(CS<'a>),
    RPL_STATSLINKINFO(CS<'a>),
    RPL_STATSCOMMANDS(CS<'a>),
    RPL_ENDOFSTATS(CS<'a>),
    RPL_STATSUPTIME(CS<'a>),
    RPL_STATSOLINE(CS<'a>),
    RPL_UMODEIS(CS<'a>),
    RPL_SERVLIST(CS<'a>),
    RPL_SERVLISTEND(CS<'a>),
    RPL_LUSERCLIENT(CS<'a>),
    RPL_LUSEROP(CS<'a>),
    RPL_LUSERUNKNOWN(CS<'a>),
    RPL_LUSERCHANNELS(CS<'a>),
    RPL_LUSERME(CS<'a>),
    RPL_ADMINME(CS<'a>),
    RPL_ADMINLOC1(CS<'a>),
    RPL_ADMINLOC2(CS<'a>),
    RPL_ADMINEMAIL(CS<'a>),
    RPL_TRYAGAIN(CS<'a>),
    ERR_NOSUCHNICK(CS<'a>),
    ERR_NOSUCHSERVER(CS<'a>),
    ERR_NOSUCHCHANNEL(CS<'a>),
    ERR_CANNOTSENDTOCHAN(CS<'a>),
    ERR_TOOMANYCHANNELS(CS<'a>),
    ERR_WASNOSUCHNICK(CS<'a>),
    ERR_TOOMANYTARGETS(CS<'a>),
    ERR_NOSUCHSERVICE(CS<'a>),
    ERR_NOORIGIN(CS<'a>),
    ERR_NORECIPIENT(CS<'a>),
    ERR_NOTEXTTOSEND(CS<'a>),
    ERR_NOTOPLEVEL(CS<'a>),
    ERR_WILDTOPLEVEL(CS<'a>),
    ERR_BADMASK(CS<'a>),
    ERR_UNKNOWNCOMMAND(CS<'a>),
    ERR_NOMOTD(CS<'a>),
    ERR_NOADMININFO(CS<'a>),
    ERR_FILEERROR(CS<'a>),
    ERR_NONICKNAMEGIVEN(CS<'a>),
    ERR_ERRONEUSNICKNAME(CS<'a>),
    ERR_NICKNAMEINUSE(CS<'a>),
    ERR_NICKCOLLISION(CS<'a>),
    ERR_UNAVAILRESOURCE(CS<'a>),
    ERR_USERNOTINCHANNEL(CS<'a>),
    ERR_NOTONCHANNEL(CS<'a>),
    ERR_USERONCHANNEL(CS<'a>),
    ERR_NOLOGIN(CS<'a>),
    ERR_SUMMONDISABLED(CS<'a>),
    ERR_USERSDISABLED(CS<'a>),
    ERR_NOTREGISTERED(CS<'a>),
    ERR_NEEDMOREPARAMS(CS<'a>),
    ERR_ALREADYREGISTRED(CS<'a>),
    ERR_NOPERMFORHOST(CS<'a>),
    ERR_PASSWDMISMATCH(CS<'a>),
    ERR_YOUREBANNEDCREEP(CS<'a>),
    ERR_YOUWILLBEBANNED(CS<'a>),
    ERR_KEYSET(CS<'a>),
    ERR_CHANNELISFULL(CS<'a>),
    ERR_UNKNOWNMODE(CS<'a>),
    ERR_INVITEONLYCHAN(CS<'a>),
    ERR_BANNEDFROMCHAN(CS<'a>),
    ERR_BADCHANNELKEY(CS<'a>),
    ERR_BADCHANMASK(CS<'a>),
    ERR_NOCHANMODES(CS<'a>),
    ERR_BANLISTFULL(CS<'a>),
    ERR_NOPRIVILEGES(CS<'a>),
    ERR_CHANOPRIVSNEEDED(CS<'a>),
    ERR_CANTKILLSERVER(CS<'a>),
    ERR_RESTRICTED(CS<'a>),
    ERR_UNIQOPPRIVSNEEDED(CS<'a>),
    ERR_NOOPERHOST(CS<'a>),
    ERR_UMODEUNKNOWNFLAG(CS<'a>),
    ERR_USERSDONTMATCH(CS<'a>),
}

Variants

RPL_WELCOME

001 RPL_WELCOME "Welcome to the Internet Relay Network !@"

RPL_YOURHOST

002 RPL_YOURHOST "Your host is , running version "

RPL_CREATED

003 RPL_CREATED "This server was created "

RPL_MYINFO

004 RPL_MYINFO " "

The server sends Replies 001 to 004 to a user upon successful registration.

RPL_BOUNCE

005 RPL_BOUNCE "Try server , port "

Sent by the server to a user to suggest an alternative server. This is often used when the connection is refused because the server is already full.

RPL_USERHOST

302 RPL_USERHOST ":*1 *( " " )"

Reply format used by USERHOST to list replies to the query list. The reply string is composed as follows:

reply = nickname [ "*" ] "=" ( "+" / "-" ) hostname

The '*' indicates whether the client has registered as an Operator. The '-' or '+' characters represent whether the client has set an AWAY message or not respectively.

RPL_ISON

303 RPL_ISON ":*1 *( " " )"

Reply format used by ISON to list replies to the query list.

RPL_AWAY

301 RPL_AWAY " :"

RPL_UNAWAY

305 RPL_UNAWAY ":You are no longer marked as being away"

RPL_NOWAWAY

306 RPL_NOWAWAY ":You have been marked as being away"

These replies are used with the AWAY command (if allowed). RPL_AWAY is sent to any client sending a PRIVMSG to a client which is away. RPL_AWAY is only sent by the server to which the client is connected. Replies RPL_UNAWAY and RPL_NOWAWAY are sent when the client removes and sets an AWAY message.

RPL_WHOISUSER

311 RPL_WHOISUSER " * :"

RPL_WHOISSERVER

312 RPL_WHOISSERVER " :"

RPL_WHOISOPERATOR

313 RPL_WHOISOPERATOR " :is an IRC operator"

RPL_WHOISIDLE

317 RPL_WHOISIDLE " :seconds idle"

RPL_ENDOFWHOIS

318 RPL_ENDOFWHOIS " :End of WHOIS list"

RPL_WHOISCHANNELS

319 RPL_WHOISCHANNELS " :*( ( "@" / "+" ) " " )"

Replies 311 - 313, 317 - 319 are all replies generated in response to a WHOIS message. Given that there are enough parameters present, the answering server MUST either formulate a reply out of the above numerics (if the query nick is found) or return an error reply. The '*' in RPL_WHOISUSER is there as the literal character and not as a wild card. For each reply set, only RPL_WHOISCHANNELS may appear more than once (for long lists of channel names). The '@' and '+' characters next to the channel name indicate whether a client is a channel operator or has been granted permission to speak on a moderated channel. The RPL_ENDOFWHOIS reply is used to mark the end of processing a WHOIS message.

RPL_WHOWASUSER

314 RPL_WHOWASUSER " * :"

RPL_ENDOFWHOWAS

369 RPL_ENDOFWHOWAS " :End of WHOWAS"

When replying to a WHOWAS message, a server MUST use the replies RPL_WHOWASUSER, RPL_WHOISSERVER or ERR_WASNOSUCHNICK for each nickname in the presented list. At the end of all reply batches, there MUST be RPL_ENDOFWHOWAS (even if there was only one reply and it was an error).

RPL_LISTSTART

321 RPL_LISTSTART Obsolete. Not used.

RPL_LIST

322 RPL_LIST " <# visible> :"

RPL_LISTEND

323 RPL_LISTEND ":End of LIST"

Replies RPL_LIST, RPL_LISTEND mark the actual replies with data and end of the server's response to a LIST command. If there are no channels available to return, only the end reply MUST be sent.

RPL_UNIQOPIS

325 RPL_UNIQOPIS " "

RPL_CHANNELMODEIS

324 RPL_CHANNELMODEIS " "

RPL_NOTOPIC

331 RPL_NOTOPIC " :No topic is set"

RPL_TOPIC

332 RPL_TOPIC " :"

When sending a TOPIC message to determine the channel topic, one of two replies is sent. If the topic is set, RPL_TOPIC is sent back else RPL_NOTOPIC.

RPL_INVITING

341 RPL_INVITING " "

Returned by the server to indicate that the attempted INVITE message was successful and is being passed onto the end client.

RPL_SUMMONING

342 RPL_SUMMONING " :Summoning user to IRC"

Returned by a server answering a SUMMON message to indicate that it is summoning that user.

RPL_INVITELIST

346 RPL_INVITELIST " "

RPL_ENDOFINVITELIST

347 RPL_ENDOFINVITELIST " :End of channel invite list"

When listing the 'invitations masks' for a given channel, a server is required to send the list back using the RPL_INVITELIST and RPL_ENDOFINVITELIST messages. A separate RPL_INVITELIST is sent for each active mask. After the masks have been listed (or if none present) a RPL_ENDOFINVITELIST MUST be sent.

RPL_EXCEPTLIST

348 RPL_EXCEPTLIST " "

RPL_ENDOFEXCEPTLIST

349 RPL_ENDOFEXCEPTLIST " :End of channel exception list"

When listing the 'exception masks' for a given channel, a server is required to send the list back using the RPL_EXCEPTLIST and RPL_ENDOFEXCEPTLIST messages. A separate RPL_EXCEPTLIST is sent for each active mask. After the masks have been listed (or if none present) a RPL_ENDOFEXCEPTLIST MUST be sent.

RPL_VERSION

351 RPL_VERSION ". :"

Reply by the server showing its version details. The is the version of the software being used (including any patchlevel revisions) and the is used to indicate if the server is running in "debug mode".

The "comments" field may contain any comments about the version or further version details.

RPL_WHOREPLY

352 RPL_WHOREPLY " ( "H" / "G" > ["*"] [ ( "@" / "+" ) ] : "

RPL_ENDOFWHO

315 RPL_ENDOFWHO " :End of WHO list"

The RPL_WHOREPLY and RPL_ENDOFWHO pair are used to answer a WHO message. The RPL_WHOREPLY is only sent if there is an appropriate match to the WHO query. If there is a list of parameters supplied with a WHO message, a RPL_ENDOFWHO MUST be sent after processing each list item with being the item.

RPL_NAMREPLY

353 RPL_NAMREPLY "( "=" / "*" / "@" ) :[ "@" / "+" ] ( " " [ "@" / "+" ] ) - "@" is used for secret channels, "" for private channels, and "=" for others (public channels).

RPL_ENDOFNAMES

366 RPL_ENDOFNAMES " :End of NAMES list"

To reply to a NAMES message, a reply pair consisting of RPL_NAMREPLY and RPL_ENDOFNAMES is sent by the server back to the client. If there is no channel found as in the query, then only RPL_ENDOFNAMES is returned. The exception to this is when a NAMES message is sent with no parameters and all visible channels and contents are sent back in a series of RPL_NAMEREPLY messages with a RPL_ENDOFNAMES to mark the end.

RPL_LINKS

364 RPL_LINKS " : "

RPL_ENDOFLINKS

365 RPL_ENDOFLINKS " :End of LINKS list"

In replying to the LINKS message, a server MUST send replies back using the RPL_LINKS numeric and mark the end of the list using an RPL_ENDOFLINKS reply.

RPL_BANLIST

367 RPL_BANLIST " "

RPL_ENDOFBANLIST

368 RPL_ENDOFBANLIST " :End of channel ban list"

When listing the active 'bans' for a given channel, a server is required to send the list back using the RPL_BANLIST and RPL_ENDOFBANLIST messages. A separate RPL_BANLIST is sent for each active banmask. After the banmasks have been listed (or if none present) a RPL_ENDOFBANLIST MUST be sent.

RPL_INFO

371 RPL_INFO ":"

RPL_ENDOFINFO

374 RPL_ENDOFINFO ":End of INFO list"

A server responding to an INFO message is required to send all its 'info' in a series of RPL_INFO messages with a RPL_ENDOFINFO reply to indicate the end of the replies.

RPL_MOTDSTART

375 RPL_MOTDSTART ":- Message of the day - "

RPL_MOTD

372 RPL_MOTD ":- "

RPL_ENDOFMOTD

376 RPL_ENDOFMOTD ":End of MOTD command"

When responding to the MOTD message and the MOTD file is found, the file is displayed line by line, with each line no longer than 80 characters, using RPL_MOTD format replies. These MUST be surrounded by a RPL_MOTDSTART (before the RPL_MOTDs) and an RPL_ENDOFMOTD (after).

RPL_YOUREOPER

381 RPL_YOUREOPER ":You are now an IRC operator"

RPL_YOUREOPER is sent back to a client which has just successfully issued an OPER message and gained operator status.

RPL_REHASHING

382 RPL_REHASHING " :Rehashing"

If the REHASH option is used and an operator sends a REHASH message, an RPL_REHASHING is sent back to the operator.

RPL_YOURESERVICE

383 RPL_YOURESERVICE "You are service "

Sent by the server to a service upon successful registration.

RPL_TIME

391 RPL_TIME " :"

When replying to the TIME message, a server MUST send the reply using the RPL_TIME format above. The string showing the time need only contain the correct day and time there. There is no further requirement for the time string.

RPL_USERSSTART

392 RPL_USERSSTART ":UserID Terminal Host"

RPL_USERS

393 RPL_USERS ": "

RPL_ENDOFUSERS

394 RPL_ENDOFUSERS ":End of users"

RPL_NOUSERS

395 RPL_NOUSERS ":Nobody logged in"

If the USERS message is handled by a server, the replies RPL_USERSTART, RPL_USERS, RPL_ENDOFUSERS and RPL_NOUSERS are used. RPL_USERSSTART MUST be sent first, following by either a sequence of RPL_USERS or a single RPL_NOUSER. Following this is RPL_ENDOFUSERS.

RPL_TRACELINK

200 RPL_TRACELINK "Link V "

RPL_TRACECONNECTING

201 RPL_TRACECONNECTING "Try. "

RPL_TRACEHANDSHAKE

202 RPL_TRACEHANDSHAKE "H.S. "

RPL_TRACEUNKNOWN

203 RPL_TRACEUNKNOWN "???? []"

RPL_TRACEOPERATOR

204 RPL_TRACEOPERATOR "Oper "

RPL_TRACEUSER

205 RPL_TRACEUSER "User "

RPL_TRACESERVER

206 RPL_TRACESERVER "Serv S C @ V"

RPL_TRACESERVICE

207 RPL_TRACESERVICE "Service "

RPL_TRACENEWTYPE

208 RPL_TRACENEWTYPE " 0 "

RPL_TRACECLASS

209 RPL_TRACECLASS "Class "

RPL_TRACERECONNECT

210 RPL_TRACERECONNECT Unused.

RPL_TRACELOG

261 RPL_TRACELOG "File "

RPL_TRACEEND

262 RPL_TRACEEND " :End of TRACE"

The RPL_TRACE* are all returned by the server in response to the TRACE message. How many are returned is dependent on the TRACE message and whether it was sent by an operator or not. There is no predefined order for which occurs first. Replies RPL_TRACEUNKNOWN, RPL_TRACECONNECTING and RPL_TRACEHANDSHAKE are all used for connections which have not been fully established and are either unknown, still attempting to connect or in the process of completing the 'server handshake'. RPL_TRACELINK is sent by any server which handles a TRACE message and has to pass it on to another server. The list of RPL_TRACELINKs sent in response to a TRACE command traversing the IRC network should reflect the actual connectivity of the servers themselves along that path.

RPL_TRACENEWTYPE is to be used for any connection which does not fit in the other categories but is being displayed anyway. RPL_TRACEEND is sent to indicate the end of the list.

RPL_STATSLINKINFO

211 RPL_STATSLINKINFO " "

reports statistics on a connection. identifies the particular connection, is the amount of data that is queued and waiting to be sent the number of messages sent, and the amount of data sent, in Kbytes. and are the equivalent of and for received data, respectively. indicates how long ago the connection was opened, in seconds.

RPL_STATSCOMMANDS

212 RPL_STATSCOMMANDS " "

reports statistics on commands usage.

RPL_ENDOFSTATS

219 RPL_ENDOFSTATS " :End of STATS report"

RPL_STATSUPTIME

242 RPL_STATSUPTIME ":Server Up %d days %d:%02d:%02d"

reports the server uptime.

RPL_STATSOLINE

243 RPL_STATSOLINE "O * "

reports the allowed hosts from where user may become IRC operators.

RPL_UMODEIS

221 RPL_UMODEIS ""

To answer a query about a client's own mode, RPL_UMODEIS is sent back.

RPL_SERVLIST

234 RPL_SERVLIST " "

RPL_SERVLISTEND

235 RPL_SERVLISTEND " :End of service listing"

When listing services in reply to a SERVLIST message, a server is required to send the list back using the RPL_SERVLIST and RPL_SERVLISTEND messages. A separate RPL_SERVLIST is sent for each service. After the services have been listed (or if none present) a RPL_SERVLISTEND MUST be sent.

RPL_LUSERCLIENT

251 RPL_LUSERCLIENT ":There are users and services on servers"

RPL_LUSEROP

252 RPL_LUSEROP " :operator(s) online"

RPL_LUSERUNKNOWN

253 RPL_LUSERUNKNOWN " :unknown connection(s)"

RPL_LUSERCHANNELS

254 RPL_LUSERCHANNELS " :channels formed"

RPL_LUSERME

255 RPL_LUSERME ":I have clients and servers"

In processing an LUSERS message, the server sends a set of replies from RPL_LUSERCLIENT, RPL_LUSEROP, RPL_USERUNKNOWN, RPL_LUSERCHANNELS and RPL_LUSERME. When replying, a server MUST send back RPL_LUSERCLIENT and RPL_LUSERME. The other replies are only sent back if a non-zero count is found for them.

RPL_ADMINME

256 RPL_ADMINME " :Administrative info"

RPL_ADMINLOC1

257 RPL_ADMINLOC1 ":"

RPL_ADMINLOC2

258 RPL_ADMINLOC2 ":"

RPL_ADMINEMAIL

259 RPL_ADMINEMAIL ":"

When replying to an ADMIN message, a server is expected to use replies RPL_ADMINME through to RPL_ADMINEMAIL and provide a text message with each. For RPL_ADMINLOC1 a description of what city, state and country the server is in is expected, followed by details of the institution (RPL_ADMINLOC2)

and finally the administrative contact for the server (an email address here is REQUIRED) in RPL_ADMINEMAIL.

RPL_TRYAGAIN

263 RPL_TRYAGAIN " :Please wait a while and try again."

When a server drops a command without processing it, it MUST use the reply RPL_TRYAGAIN to inform the originating client.

ERR_NOSUCHNICK

401 ERR_NOSUCHNICK " :No such nick/channel"

Used to indicate the nickname parameter supplied to a command is currently unused.

ERR_NOSUCHSERVER

402 ERR_NOSUCHSERVER " :No such server"

Used to indicate the server name given currently does not exist.

ERR_NOSUCHCHANNEL

403 ERR_NOSUCHCHANNEL " :No such channel"

Used to indicate the given channel name is invalid.

ERR_CANNOTSENDTOCHAN

404 ERR_CANNOTSENDTOCHAN " :Cannot send to channel"

Sent to a user who is either (a) not on a channel which is mode +n or (b) not a chanop (or mode +v) on a channel which has mode +m set or where the user is banned and is trying to send a PRIVMSG message to that channel.

ERR_TOOMANYCHANNELS

405 ERR_TOOMANYCHANNELS " :You have joined too many channels"

Sent to a user when they have joined the maximum number of allowed channels and they try to join another channel.

ERR_WASNOSUCHNICK

406 ERR_WASNOSUCHNICK " :There was no such nickname"

Returned by WHOWAS to indicate there is no history information for that nickname.

ERR_TOOMANYTARGETS

407 ERR_TOOMANYTARGETS " : recipients. "

Returned to a client which is attempting to send a PRIVMSG/NOTICE using the user@host destination format and for a user@host which has several occurrences.

Returned to a client which trying to send a PRIVMSG/NOTICE to too many recipients.

Returned to a client which is attempting to JOIN a safe channel using the shortname when there are more than one such channel.

ERR_NOSUCHSERVICE

408 ERR_NOSUCHSERVICE " :No such service"

Returned to a client which is attempting to send a SQUERY to a service which does not exist.

ERR_NOORIGIN

409 ERR_NOORIGIN ":No origin specified"

PING or PONG message missing the originator parameter.

ERR_NORECIPIENT

411 ERR_NORECIPIENT ":No recipient given ()"

ERR_NOTEXTTOSEND

412 ERR_NOTEXTTOSEND ":No text to send"

ERR_NOTOPLEVEL

413 ERR_NOTOPLEVEL " :No toplevel domain specified"

ERR_WILDTOPLEVEL

414 ERR_WILDTOPLEVEL " :Wildcard in toplevel domain"

ERR_BADMASK

415 ERR_BADMASK " :Bad Server/host mask"

412 - 415 are returned by PRIVMSG to indicate that the message wasn't delivered for some reason. ERR_NOTOPLEVEL and ERR_WILDTOPLEVEL are errors that are returned when an invalid use of "PRIVMSG $" or "PRIVMSG #" is attempted.

ERR_UNKNOWNCOMMAND

421 ERR_UNKNOWNCOMMAND " :Unknown command"

Returned to a registered client to indicate that the command sent is unknown by the server.

ERR_NOMOTD

422 ERR_NOMOTD ":MOTD File is missing"

Server's MOTD file could not be opened by the server.

ERR_NOADMININFO

423 ERR_NOADMININFO " :No administrative info available"

Returned by a server in response to an ADMIN message when there is an error in finding the appropriate information.

ERR_FILEERROR

424 ERR_FILEERROR ":File error doing on "

Generic error message used to report a failed file operation during the processing of a message.

ERR_NONICKNAMEGIVEN

431 ERR_NONICKNAMEGIVEN ":No nickname given"

Returned when a nickname parameter expected for a command and isn't found.

ERR_ERRONEUSNICKNAME

432 ERR_ERRONEUSNICKNAME " :Erroneous nickname"

Returned after receiving a NICK message which contains characters which do not fall in the defined set. See section 2.3.1 for details on valid nicknames.

ERR_NICKNAMEINUSE

433 ERR_NICKNAMEINUSE " :Nickname is already in use"

Returned when a NICK message is processed that results in an attempt to change to a currently existing nickname.

ERR_NICKCOLLISION

436 ERR_NICKCOLLISION " :Nickname collision KILL from @"

Returned by a server to a client when it detects a nickname collision (registered of a NICK that already exists by another server).

ERR_UNAVAILRESOURCE

437 ERR_UNAVAILRESOURCE " :Nick/channel is temporarily unavailable"

Returned by a server to a user trying to join a channel currently blocked by the channel delay mechanism.

Returned by a server to a user trying to change nickname when the desired nickname is blocked by the nick delay mechanism.

ERR_USERNOTINCHANNEL

441 ERR_USERNOTINCHANNEL " :They aren't on that channel"

Returned by the server to indicate that the target user of the command is not on the given channel.

ERR_NOTONCHANNEL

442 ERR_NOTONCHANNEL " :You're not on that channel"

Returned by the server whenever a client tries to perform a channel affecting command for which the client isn't a member.

ERR_USERONCHANNEL

443 ERR_USERONCHANNEL " :is already on channel"

Returned when a client tries to invite a user to a channel they are already on.

ERR_NOLOGIN

444 ERR_NOLOGIN " :User not logged in"

Returned by the summon after a SUMMON command for a user was unable to be performed since they were not logged in.

ERR_SUMMONDISABLED

445 ERR_SUMMONDISABLED ":SUMMON has been disabled"

Returned as a response to the SUMMON command. MUST be returned by any server which doesn't implement it.

ERR_USERSDISABLED

446 ERR_USERSDISABLED ":USERS has been disabled"

Returned as a response to the USERS command. MUST be returned by any server which does not implement it.

ERR_NOTREGISTERED

451 ERR_NOTREGISTERED ":You have not registered"

Returned by the server to indicate that the client MUST be registered before the server will allow it to be parsed in detail.

ERR_NEEDMOREPARAMS

461 ERR_NEEDMOREPARAMS " :Not enough parameters"

Returned by the server by numerous commands to indicate to the client that it didn't supply enough parameters.

ERR_ALREADYREGISTRED

462 ERR_ALREADYREGISTRED ":Unauthorized command (already registered)"

Returned by the server to any link which tries to change part of the registered details (such as password or user details from second USER message).

ERR_NOPERMFORHOST

463 ERR_NOPERMFORHOST ":Your host isn't among the privileged"

Returned to a client which attempts to register with a server which does not been setup to allow connections from the host the attempted connection is tried.

ERR_PASSWDMISMATCH

464 ERR_PASSWDMISMATCH ":Password incorrect"

Returned to indicate a failed attempt at registering a connection for which a password was required and was either not given or incorrect.

ERR_YOUREBANNEDCREEP

465 ERR_YOUREBANNEDCREEP ":You are banned from this server"

Returned after an attempt to connect and register yourself with a server which has been setup to explicitly deny connections to you.

ERR_YOUWILLBEBANNED

466 ERR_YOUWILLBEBANNED

Sent by a server to a user to inform that access to the server will soon be denied.

ERR_KEYSET

467 ERR_KEYSET " :Channel key already set"

ERR_CHANNELISFULL

471 ERR_CHANNELISFULL " :Cannot join channel (+l)"

ERR_UNKNOWNMODE

472 ERR_UNKNOWNMODE " :is unknown mode char to me for "

ERR_INVITEONLYCHAN

473 ERR_INVITEONLYCHAN " :Cannot join channel (+i)"

ERR_BANNEDFROMCHAN

474 ERR_BANNEDFROMCHAN " :Cannot join channel (+b)"

ERR_BADCHANNELKEY

475 ERR_BADCHANNELKEY " :Cannot join channel (+k)"

ERR_BADCHANMASK

476 ERR_BADCHANMASK " :Bad Channel Mask"

ERR_NOCHANMODES

477 ERR_NOCHANMODES " :Channel doesn't support modes"

ERR_BANLISTFULL

478 ERR_BANLISTFULL " :Channel list is full"

ERR_NOPRIVILEGES

481 ERR_NOPRIVILEGES ":Permission Denied- You're not an IRC operator"

Any command requiring operator privileges to operate MUST return this error to indicate the attempt was unsuccessful.

ERR_CHANOPRIVSNEEDED

482 ERR_CHANOPRIVSNEEDED " :You're not channel operator"

Any command requiring 'chanop' privileges (such as MODE messages) MUST return this error if the client making the attempt is not a chanop on the specified channel.

ERR_CANTKILLSERVER

483 ERR_CANTKILLSERVER ":You can't kill a server!"

Any attempts to use the KILL command on a server are to be refused and this error returned directly to the client.

ERR_RESTRICTED

484 ERR_RESTRICTED ":Your connection is restricted!"

Sent by the server to a user upon connection to indicate the restricted nature of the connection (user mode "+r").

ERR_UNIQOPPRIVSNEEDED

485 ERR_UNIQOPPRIVSNEEDED ":You're not the original channel operator"

Any MODE requiring "channel creator" privileges MUST return this error if the client making the attempt is not a chanop on the specified channel.

ERR_NOOPERHOST

491 ERR_NOOPERHOST ":No O-lines for your host"

If a client sends an OPER message and the server has not been configured to allow connections from the client's host as an operator, this error MUST be returned.

ERR_UMODEUNKNOWNFLAG

501 ERR_UMODEUNKNOWNFLAG ":Unknown MODE flag"

Returned by the server to indicate that a MODE message was sent with a nickname parameter and that the a mode flag sent was not recognized.

ERR_USERSDONTMATCH

502 ERR_USERSDONTMATCH ":Cannot change mode for other users"

Error sent to any user trying to view or change the user mode for a user other than themselves.

Methods

impl<'a> Reply<'a>

fn from_message(msg: &'a Message) -> Option<Reply<'a>>

fn to_message(&'a self) -> Message

fn to_static(&self) -> Reply<'static>